refactor rppg pulse analysis
